refactor(git-id-switcher): split htmlTemplates.ts into a directory and tighten the shell trust boundary Break the monolithic 462-line `src/ui/htmlTemplates.ts` into a `htmlTemplates/` package with per-concern files: - `shell.ts`  — CSP validators, getBaseStyles, buildHtmlShell (the   `<!DOCTYPE>…</head>` skeleton), shared focus-visible rule helpers,   and the SSOT constants (body-class allowlist, GitHub README URL). - `document.ts` / `loading.ts` / `error.ts` — each template owns only   its body markup and template-specific CSS, with `ERROR_MESSAGES`   promoted to a module constant and `ErrorType` colocated with the   only consumer. - `index.ts` — pure re-export barrel so existing `./htmlTemplates`   imports resolve unchanged. Additional hardening folded into the same change: - `buildHtmlShell` now owns the `getBaseStyles() + extraStyles`   concatenation, so templates cannot drift on the base layer. - `HtmlShellOptions.body` is typed as `SanitizedHtml` (brand), making   the trust boundary symmetric with `buildDocumentHtml#content` — a   raw `string` body now fails at compile time. - `extraStyles` gets a defense-in-depth `</style` breakout guard so a   future dynamic caller cannot escape the `<style>` raw-text element. - Shared focus-visible rule extracted to `getFocusVisibleRule` /   `getFocusVisibleForcedColorsRule` helpers, deduplicating the block   between document and error templates. - Loading template's bare `p {}` selector is scoped under   `body.gis-loading p` to match the other templates. Tests added: barrel re-export surface (including `BodyClass` / `buildHtmlShell`), `extraStyles` breakout guard positive+negative, and the loading scope fix. Coverage for `src/ui/htmlTemplates/` stays at 100% statements / 100% branches / 100% functions / 100% lines. Signed-off-by: Null;Variant <null@nullvariant.com>
🖥️ IDE: [VS Code](https://code.visualstudio.com/)
🔌 Extension: [Claude Code](https://claude.ai/download)

Model-Raw: claude-opus-4-6
